+++ test/store/settings_test.dart
@@ -3,7 +3,7 @@ import 'package:polka_wallet/common/consts/settings.dart';
 import 'package:polka_wallet/store/app.dart';
 import 'package:polka_wallet/store/settings.dart';
 
-import 'localStorage_mock.dart';
+import '../mocks/localStorage_mock.dart';
 
 void main() {
   TestWidgetsFlutterBinding.ensureInitialized();
